+++ src/Retcon.php
@@ -9,6 +9,7 @@ use craft\web\twig\variables\CraftVariable;
 use mmikkel\retcon\services\RetconService;
 use mmikkel\retcon\twigextensions\RetconTwigExtension;
 use mmikkel\retcon\models\RetconSettings;
+use mmikkel\retcon\variables\RetconVariable;
 
 use yii\base\Event;
 
@@ -36,14 +37,14 @@ class Retcon extends Plugin
         // Add in our Twig extensions
         Craft::$app->getView()->registerTwigExtension(new RetconTwigExtension());
 
-        // Register our variables
+        // Register our variables (this is deprecated in Retcon 3.0.0 and will be removed in Retcon 4)
         Event::on(
             CraftVariable::class,
             CraftVariable::EVENT_INIT,
             function (Event $event) {
                 /** @var CraftVariable $variable */
                 $variable = $event->sender;
-                $variable->set('retcon', RetconService::class);
+                $variable->set('retcon', RetconVariable::class);
             }
         );
     }